requirements.txt
Flask
mysql-connector-python
werkzeug
flask_socketio
python-dotenv
APScheduler